+++ pricing/src/epilot/availability_api.py
@@ -21,13 +21,13 @@ class AvailabilityAPI:
         
         url = base_url + '/v1/public/availability:check'
         headers = utils.get_headers(request)
-        req_content_type, data, form = utils.serialize_request_body(request, "availability_check_params", 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, "availability_check_params", False, False, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         if data is None and form is None:
             raise Exception('request body is required')
         headers['Accept'] = 'application/json'
-        headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
+        headers['user-agent'] = self.sdk_configuration.user_agent
         
         client = self.sdk_configuration.security_client