+++ pricing/src/epilot/models/shared/catalogsearch.py
@@ -19,7 +19,7 @@ class CatalogSearch:
     r"""Availability filters dimensions"""
     from_: Optional[float] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('from'), 'exclude': lambda f: f is None }})
     r"""The index from which to query, used for pagination purposes. Defaults to 0"""
-    hydrate: Optional[bool] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('hydrate'), 'exclude': lambda f: f is None }})
+    hydrate: Optional[bool] = dataclasses.field(default=False, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('hydrate'), 'exclude': lambda f: f is None }})
     r"""When true, enables entity hydration to resolve nested $relation references in-place."""
     size: Optional[float] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('size'), 'exclude': lambda f: f is None }})
     r"""The max size of the response, defaults to 2000."""